+++ wazo_auth/plugins/http/user_registration/plugin.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+# Copyright 2017 The Wazo Authors  (see the AUTHORS file)
+# SPDX-License-Identifier: GPL-3.0+
+
+from . import http
+
+
+class Plugin(object):
+
+    def load(self, dependencies):
+        api = dependencies['api']
+        args = (dependencies['user_service'],)
+
+        api.add_resource(http.Register, '/users/register', resource_class_args=args)